# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.1",
    "title": "Car Racing Game",
    "description": "Apis for car racing game"
  },
  "host": "localhost:3500",
  "basePath": "",
...
  "schemes": [],
  "securityDefinitions": {},
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/api/races/{id}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
...
    "/api/races/{id}/start": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/races/{id}/accelerate": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    }
  },
...
}